requirements.txt
Flask
numpy
pandas
Flask-Cors
openpyxl
reportlab
matplotlib
requests
python-bidi
